# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation pour l'API de gestion des catégories de votre application.",
        "contact": {
            "email": "votre_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/games": {
            "get": {
                "summary": "Récupérer toutes les games",
                "operationId": "93736f40a135111a310e4b25ad61f30b",
                "responses": {
                    "200": {
                        "description": "Liste des games"
                    }
                }
            },
            "post": {
                "summary": "Crée nouveau jeu",
                "operationId": "b8b9557618eb0f7b3650bdd4ab3a6487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "groupe_id",
                                    "candidat",
                                    "phone",
                                    "gift_id"
                                ],
                                "properties": {
                                    "groupe_id": {
                                        "type": "text"
                                    },
                                    "candidat": {
                                        "type": "text"
                                    },
                                    "phone": {
                                        "type": "text"
                                    },
                                    "gift_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jeu est crée"
                    }
                }
            }
        },
        "/api/games/{id}": {
            "put": {
                "summary": "Modifier Jeu",
                "operationId": "7a7a51e92437ab9405a4bd9153eacec7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "groupe_id",
                                    "candidat",
                                    "phone",
                                    "gift_id"
                                ],
                                "properties": {
                                    "groupe_id": {
                                        "type": "text"
                                    },
                                    "candidat": {
                                        "type": "text"
                                    },
                                    "phone": {
                                        "type": "text"
                                    },
                                    "gift_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jeu est mise à jour"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer le Jeu",
                "operationId": "1b52414d7af18b35901add3451cba91c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jeu est supprimé"
                    }
                }
            }
        },
        "/api/groupes/{groupe}/games": {
            "get": {
                "summary": "Récupérer tous les joueurs d'un groupe",
                "description": "Récupère la liste de tous les joueurs associés à un groupe spécifique.",
                "operationId": "4d7ebdc9f0b83e66aa9aed8598360de9",
                "parameters": [
                    {
                        "name": "groupe",
                        "in": "path",
                        "description": "ID du groupe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des joueurs du groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 101
                                            },
                                            "nom": {
                                                "type": "string",
                                                "example": "Nom du joueur"
                                            },
                                            "groupe_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Groupe non trouvé"
                    }
                }
            }
        },
        "/api/gifts": {
            "get": {
                "summary": "Récupérer toutes les cadeaux",
                "operationId": "4befadf53537a7ab77d7b731ecafcfee",
                "responses": {
                    "200": {
                        "description": "Liste des cadeaux"
                    }
                }
            },
            "post": {
                "summary": "Ajouter un cadeau",
                "operationId": "84cc149646671c5808f7cecd619bbd0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cadeau est crée"
                    }
                }
            }
        },
        "/api/gifts/{id}": {
            "put": {
                "summary": "Mettre à jour le cadeau",
                "operationId": "3f9d080ff64fcd36e707fa4f8f675be8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cadeau est mise à jour"
                    },
                    "404": {
                        "description": "Gift not found"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer le cadeau",
                "operationId": "d36bc890268318f32be19cc32e70cd46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cadeau est supprimé"
                    }
                }
            }
        },
...
    },
...
}